Move DAOs into package representing their backing database
