+++ src/main/java/com/dongtronic/diabot/data/redis/NightscoutDTO.kt
@@ -1,5 +1,6 @@
-package com.dongtronic.diabot.data
+package com.dongtronic.diabot.data.redis
 
+import com.dongtronic.diabot.data.ConversionDTO
 import java.time.ZonedDateTime
 
 class NightscoutDTO {
