+++ src/main/java/com/dongtronic/diabot/data/redis/RuleDTO.kt
@@ -1,4 +1,4 @@
-package com.dongtronic.diabot.data
+package com.dongtronic.diabot.data.redis
 
 
 class RuleDTO(var id: String, var title: String, var text: String, var messageId: String)
