+++ src/main/java/com/dongtronic/diabot/platforms/discord/commands/admin/username/AdminUsernamePatternCommand.kt
@@ -1,6 +1,6 @@
 package com.dongtronic.diabot.platforms.discord.commands.admin.username
 
-import com.dongtronic.diabot.data.AdminDAO
+import com.dongtronic.diabot.data.redis.AdminDAO
 import com.dongtronic.diabot.platforms.discord.commands.DiscordCommand
 import com.dongtronic.diabot.util.Logger
 import com.jagrosh.jdautilities.command.Command
