+++ src/main/java/com/dongtronic/diabot/platforms/discord/commands/nightscout/NightscoutAdminSimpleDeleteCommand.kt
@@ -1,6 +1,6 @@
 package com.dongtronic.diabot.platforms.discord.commands.nightscout
 
-import com.dongtronic.diabot.data.NightscoutDAO
+import com.dongtronic.diabot.data.redis.NightscoutDAO
 import com.dongtronic.diabot.platforms.discord.commands.DiscordCommand
 import com.dongtronic.diabot.platforms.discord.utils.CommandUtils
 import com.dongtronic.diabot.util.Logger
